Added nest syntax: sweep()(sweep())
