+++ qcodes/sweep/__init__.py
@@ -1,2 +1,2 @@
-from .functions import sweep, nest, chain, szip
+from .sweep import sweep, nest, chain, szip, measurement, setter
 from .sweep_measurement import SweepMeasurement
