+++ qcodes/sweep/sweep.py
@@ -1,6 +1,8 @@
 import itertools
 import numpy as np
 
+import qcodes
+
 
 class ParametersTable:
     def __init__(self, table_list=None, dependent_parameters=None, independent_parameters=None):
@@ -77,6 +79,23 @@ def setter(param_list):
     return decorator
 
 
+def wrap_objects(*objects, repeat=False):
+
+    def wrapper(obj):
+        if isinstance(obj, qcodes.Parameter):
+            new_obj = ParameterWrapper(obj, repeat=repeat)
+        elif isinstance(obj, BaseSweepObject):
+            new_obj = obj
+        elif callable(obj):
+            new_obj = FunctionWrapper(obj, repeat=repeat)
+        else:
+            raise ValueError("Do not know how to wrap instance of ", type(obj))
+
+        return new_obj
+
+    return [wrapper(obj) for obj in objects]
+
+
 class BaseSweepObject:
     """
     A sweep object is defined as follows:
@@ -106,6 +125,9 @@ class BaseSweepObject:
         if self._param_setter is None:
             self._param_setter = self._setter_factory()
         return next(self._param_setter)
+
+    def __call__(self, *sweep_objects):
+        return Nest([self, Chain(wrap_objects(*sweep_objects))])
 
     @property
     def parameter_table(self):
@@ -306,3 +328,61 @@ class FunctionWrapper(BaseSweepObject):
         while not stop:
             yield self._measure_function()
             stop = not self._repeat
+
+
+def sweep(obj, sweep_points):
+    """
+    A convenience function to create a 1D sweep object
+
+    Parameters
+    ----------
+    obj: qcodes.StandardParameter or callable
+        If callable, it shall be a callable of three parameters: station, namespace, set_value and shall return a
+        dictionary
+    sweep_points: iterable or callable returning a iterable
+        If callable, it shall be a callable of two parameters: station, namespace and shall return an iterable
+
+    Returns
+    -------
+    FunctionSweep or ParameterSweep
+    """
+
+    if not callable(sweep_points):
+        point_function = lambda: sweep_points
+    else:
+        point_function = sweep_points
+
+    if not isinstance(obj, qcodes.Parameter):
+        if not callable(obj):
+            raise ValueError("The object to sweep over needs to either be a QCoDeS parameter or a function")
+
+        return FunctionSweep(obj, point_function)
+    else:
+        return ParameterSweep(obj, point_function)
+
+
+def nest(*objects):
+    return Nest(wrap_objects(*objects))
+
+
+def chain(*objects):
+    return Chain(wrap_objects(*objects))
+
+
+def szip(*objects):
+    """
+    A plausible scenario for using szip is the following:
+
+    >>> szip(therometer.t, sweep(source.voltage, [0, 1, 2]))
+
+    The idea is to measure the temperature *before* going to each voltage set point. The parameter "thermometer.t" needs
+    to be wrapped by the ParameterWrapper in such a way that the get method of the parameter is called repeatedly. An
+    infinite loop is prevented because the sweep object "sweep(source.voltage, [0, 1, 2])"  has a finite length and the
+    Zip operator loops until the shortest sweep object is exhausted
+    """
+    repeat = False
+    if any([isinstance(i, BaseSweepObject) for i in objects]):  # If any of the objects is a sweep object it (probably)
+        # has a finite length and therefore wrapping functions and parameters with repeat=True is save. These parameters
+        # and functions will be called as often as the length of the sweep object.
+        repeat = True
+    return Zip(wrap_objects(*objects, repeat=repeat))
